touch_event_srv: add a generic keyboard event
the keyboard driver we're using implemented a
wide veriaty of events.
while the old keyboard callback, were expecting string only,
now we will be able to receive and parse int-s.
and pass the relevant key_event to the driver,
for example, the syntax we expect inorder to make volume changes:

 - "v 94\0" - volume up
 - "v 93\0" - volume down

Change-Id: I6c321ef8948df908ec627d7a1d7ebd788f9ddc2c
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>